Move Rune Bridge api/abi inside RuneBridgeDialog
